astgen-regression/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astgen-regression"
version = "1.0.0"
description = "Config-driven regression testing framework for language-specific astgen variants"
authors = [{name = "Joern Team"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "PyYAML>=6.0",
    "Jinja2>=3.1",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
astgen-regression = "astgen_regression.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
astgen_regression = ["templates/*.j2"]
